requirements.txt
djangorestframework
PyJWT
cryptography
psycopg2-binary
Django
bcrypt
